# JsonGenerator/endpoint-server/openapi/swaggerInit.yaml
...
openapi: 3.1.0
info:
  title: GeoRDFBench Framework JSON Specification API with Swagger
  version: 2.0.0
  description: A cRud API application made with Express and documented with Swagger
  license:
    name: Apache-2.0
    url: https://spdx.org/licenses/Apache-2.0.html
  contact:
    name: Theofilos Ioannidis
    url: http://users.uoa.gr/~tioannid/
    email: tioannid@di.uoa.gr
servers:
  - url: http://localhost:{port}/
    description: Development server
    variables:
      port:
        enum:
          - "5000"
          - "3000"
        default: "5000"
  - url: "{scheme}://geordfbench.di.uoa.gr/jsonapi/"
    description: Production server
    variables:
      scheme:
        enum:
          - http
          - https
        default: http
paths:
...
  "/reportspecs/{spec}":
    get:
      summary: Get the report (logging) by specification name
      tags:
        - Report (Logging) Specifications
      parameters:
        - in: path
          name: spec
          schema:
            type: string
          required: true
          description: The report (logging) specification name
      responses:
        "200":
          description:
            The specification for the report (logging) specification name
            provided
          contens:
            application/json:
              schema:
                "$ref": "#/components/schemas/ReportSpecification"
        "404":
          description: The report (logging) specification name was not found
  "/reportspecs/{newspec}":
    post:
      summary: Create the report (logging) by specification name
      requestBody:
        description: Specify a **JSON** string
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ReportSpecification"
      tags:
        - Report (Logging) Specifications
      parameters:
        - in: path
          name: newspec
          schema:
            type: string
          required: true
          description: The report (logging) specification name
      responses:
        "201":
          description: The report (logging) specification created
          contens:
            application/json:
              schema:
                "$ref": "#/components/schemas/ReportSpecification"
        "400":
          description: The field noQueryResultToReport is either missing or not integer!
        "404":
          description: The report (logging) specification name already exists!
  "/reportspecs/{existingspec}":
    delete:
      summary: Delete an existing report (logging) specification by name
      tags:
        - Report (Logging) Specifications
      parameters:
        - in: path
          name: existingspec
          schema:
            type: string
          required: true
          description: The report (logging) specification name
      responses:
        "204":
          description: The report (logging) specification was deleted
        "404":
          description: The report (logging) specification name was not found
...
  "/reportsources/{spec}":
    get:
      summary: Get the report source by specification name
      tags:
        - Report Sources
      parameters:
        - in: path
          name: spec
          schema:
            type: string
          required: true
          description: The report source specification name
      responses:
        "200":
          "404":
            description: The report source specification name was not found
          description:
            The specification for the report source specification name
            provided
          contens:
            application/json:
              schema:
                oneOf:
                  - "$ref": "#/components/schemas/JDBCReportSource"
                  - "$ref": "#/components/schemas/EmbeddedJDBCReportSource"
  "/reportsources/{newspec}":
    post:
      summary: Create the report source by specification name
      requestBody:
        description: Specify a **JSON** string
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - "$ref": "#/components/schemas/JDBCReportSource"
                - "$ref": "#/components/schemas/EmbeddedJDBCReportSource"
      tags:
        - Report Sources
      parameters:
        - in: path
          name: newspec
          schema:
            type: string
          required: true
          description: The report source specification name
      responses:
        "201":
          description: The report source specification was created
          contens:
            application/json:
              schema:
                oneOf:
                  - "$ref": "#/components/schemas/JDBCReportSource"
                  - "$ref": "#/components/schemas/EmbeddedJDBCReportSource"
        "400":
          description: Base Report Source fields are missing!
        "403":
          description: Embedded JDBC or JDBC Report Source specific fields are missing!
        "404":
          description: The report source specification name already exists!
  "/reportsources/{existingspec}":
    delete:
      summary: Delete an existing report source by specification name
      tags:
        - Report Sources
      parameters:
        - in: path
          name: existingspec
          schema:
            type: string
          required: true
          description: The report source specification name
      responses:
        "204":
          description: The report source specification was deleted
        "404":
          description: The report source specification name was not found
...
components:
  schemas:
...
    ReportSpecification:
      type: array
      required:
        - noQueryResultToReport
      properties:
        noQueryResultToReport:
          type: number
          format: int32
          description: The number of results to display for each query execution
      example:
        noQueryResultToReport: 3
...
    BaseReportSource:
      type: object
      required:
        - classname
        - driver
        - database
        - user
        - password
      properties:
        classname:
          type: string
          description: The java full class name
        driver:
          type: string
          enum:
            - postgresql
            - h2
        database:
          type: string
          description: The database name
        user:
          type: string
          description: The database user name
        password:
          type: string
          description: The database user password
    JDBCReportSource:
      type: object
      allOf:
        - $ref: "#/components/schemas/BaseReportSource"
          type: object
          required:
            - hostname
            - althostname
            - port
          properties:
            hostname:
              type: string
              description: The PostgreSQL server host name or IPv4
            althostname:
              type: string
              description: The alternative PostgreSQL server host name or IPv4
            port:
              type: number
              format: int32
              description: The PostgreSQL server's listening port
      example:
        classname: gr.uoa.di.rdf.geordfbench.runtime.reportsource.impl.PostgreSQLRepSrc
        driver: postgresql
        hostname: 10.0.2.15
        althostname: localhost
        port: 5432
        database: geordfbench
        user: geordfbench
        password: geordfbench
    EmbeddedJDBCReportSource:
      allOf:
        - $ref: "#/components/schemas/BaseReportSource"
          type: object
          required:
            - relativeBaseDir
          properties:
            relativeBaseDir:
              type: string
              description: The database path relative to the host's sourceFileDir
      example:
        classname: gr.uoa.di.rdf.geordfbench.runtime.reportsource.impl.H2EmbeddedRepSrc
        driver: h2
        relativeBaseDir: "../scripts/h2embeddedreportsource"
        database: geordfbench
        user: sa
        password: ""
...